+++ P2/SUC/Ejercicio_3_SUC.c
